ejercicios

+++ ejercicio4.py
@@ -1,5 +1,5 @@
-num1 = float(input("Ingrese el primer número: "))
-num2 = float(input("Ingrese el segundo número: "))
+num1 = int(input("Ingrese el primer número: "))
+num2 = int(input("Ingrese el segundo número: "))
 print("Suma:", num1 + num2)
 print("Resta:", num1 - num2)
 print("Multiplicación:", num1 * num2)
@@ -7,3 +7,4 @@ if num2 != 0:
     print("División:", num1 / num2)
 else:
     print("División: No es posible dividir entre 0")
+
